+++ PSSC/Models/Deanship/Deanship.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Models.Deanship
+{
+    public class Deanship
+    {
+        public Deanship()
+        {
+
+        }
+    }
+}
