+++ PSSC/Repositories/Deanship/RepositoryDeanship.cs
@@ -3,14 +3,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace Repositories.Professor
+namespace Repositories.Deanship
 {
-    public class ProfessorRepository : IRepositoryProfessor
+    public class RepositoryDeanship : IRepositoryDeanship
     {
         private static List<Models.Professor.Professor> _professor = new List<Models.Professor.Professor>();
+        private static List<Models.Student.Student> _student = new List<Models.Student.Student>();
 
-
-        public ProfessorRepository()
+        public RepositoryDeanship()
         {
         }
 
@@ -36,6 +36,28 @@ namespace Repositories.Professor
         {
             return _professor.FirstOrDefault(d => d.Name.Text == name);
         }
+        public void AdaugaStudent(Models.Student.Student student)
+        {
+            var result = _student.FirstOrDefault(d => d.Equals(student));
 
+            if (result != null) throw new DuplicateWaitObjectException();
+
+            _student.Add(student);
+            Console.WriteLine("Un nou Student a fost adaugat.");
+        }
+
+
+        public void ActualizeazaStudent(Models.Student.Student student)
+        {
+
+            Console.WriteLine("Modificarile au fost salvate.");
+        }
+
+
+        public Models.Student.Student GasesteStudentDupaNume(string name)
+        {
+            return _student.FirstOrDefault(d => d.Name.Text == name);
+        }
+       
     }
 }
