+++ PSSC/Repositories/Repositories.csproj
@@ -43,16 +43,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Professor\IRepositoryProfessor.cs" />
-    <Compile Include="Professor\RepositoryProfessor.cs" />
-    <Compile Include="Student\IRepositoryStudent.cs" />
-    <Compile Include="Student\RepositoryStudent.cs" />
+    <Compile Include="Deanship\IRepositoryDeanship.cs" />
+    <Compile Include="Deanship\RepositoryDeanship.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Subject\IRepositorySubject.cs" />
     <Compile Include="Subject\RepositorySubject.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <WCFMetadata Include="Service References\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
