+++ PSSC/Repositories/Subject/IRepositorySubject.cs
@@ -4,9 +4,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Repositories
+namespace Repositories.Subject
 {
     public interface IRepositorySubject
     {
+        void AdaugaSubject(Models.Subject.Subject subject);
+        void ActualizeazaSubject(Models.Subject.Subject subject);
+        Models.Subject.Subject subject(string name);
     }
 }
